Enum db_traits::errors::DbErrorKind

pub enum DbErrorKind {
    NotYetImplemented(String),
    BadValuePair(String, ValueType),
    BadSQLValuePair(Value, i32),
    BadBootstrapDefinition(String),
    BadSchemaAssertion(String),
    UnrecognizedIdent(String),
    UnrecognizedEntid(Entid),
    UnallocatedEntid(Entid),
    UnknownAttribute(Entid),
    CannotCacheNonUniqueAttributeInReverse(Entid),
    SchemaAlterationFailed(String),
    SchemaConstraintViolation(SchemaConstraintViolation),
    InputError(InputError),
    WrongTypeValueForFtsAssertion,
    CacheUpdateFailed,
    CouldNotSetVersionPragma,
    CouldNotGetVersionPragma,
    CouldNotSearch,
    TxInsertFailedToAddMissingDatoms,
    TxInsertFailedToRetractDatoms,
    DatomsUpdateFailedToRetract,
    DatomsUpdateFailedToAdd,
    FailedToCreateTempTables,
    NonFtsInsertionIntoTempSearchTableFailed,
    FtsInsertionFailed,
    FtsInsertionIntoTempSearchTableFailed,
    FtsFailedToDropSearchIds,
    FailedToUpdatePartitionMap,
    TimelinesMixed,
    TimelinesMoveToNonEmpty,
    TimelinesInvalidRange,
    RusqliteError(String),
}

Variants
NotYetImplemented(String)

We're just not done yet. Message that the feature is recognized but not yet implemented.

BadValuePair(String, ValueType)

We've been given a value that isn't the correct Mentat type.

BadSQLValuePair(Value, i32)

We've got corrupt data in the SQL store: a value and value_type_tag don't line up. TODO _1.data_type()

BadBootstrapDefinition(String)

A bootstrap definition couldn't be parsed or installed. This is a programmer error, not a runtime error.

BadSchemaAssertion(String)

A schema assertion couldn't be parsed.

UnrecognizedIdent(String)

An ident->entid mapping failed.

UnrecognizedEntid(Entid)

An entid->ident mapping failed.

UnallocatedEntid(Entid)

Tried to transact an entid that isn't allocated.

UnknownAttribute(Entid)CannotCacheNonUniqueAttributeInReverse(Entid)SchemaAlterationFailed(String)SchemaConstraintViolation(SchemaConstraintViolation)

A transaction tried to violate a constraint of the schema of the Mentat store.

InputError(InputError)

The transaction was malformed in some way (that was not recognized at parse time; for example, in a way that is schema-dependent).

WrongTypeValueForFtsAssertionCacheUpdateFailedCouldNotSetVersionPragmaCouldNotGetVersionPragmaCouldNotSearchTxInsertFailedToAddMissingDatomsTxInsertFailedToRetractDatomsDatomsUpdateFailedToRetractDatomsUpdateFailedToAddFailedToCreateTempTablesNonFtsInsertionIntoTempSearchTableFailedFtsInsertionFailedFtsInsertionIntoTempSearchTableFailedFtsFailedToDropSearchIdsFailedToUpdatePartitionMapTimelinesMixedTimelinesMoveToNonEmptyTimelinesInvalidRangeRusqliteError(String)