Enum db_traits::errors::SchemaConstraintViolation

pub enum SchemaConstraintViolation {
    ConflictingUpserts {
        conflicting_upserts: BTreeMap<TempId, BTreeSet<KnownEntid>>,
    },
    TypeDisagreements {
        conflicting_datoms: BTreeMap<(Entid, Entid, TypedValue), ValueType>,
    },
    CardinalityConflicts {
        conflicts: Vec<CardinalityConflict>,
    },
}

Variants
ConflictingUpserts

A transaction tried to assert datoms where one tempid upserts to two (or more) distinct entids.

Fields of ConflictingUpserts

conflicting_upserts: BTreeMap<TempId, BTreeSet<KnownEntid>>

A map from tempid to the entids it would upsert to.

In the future, we might even be able to attribute the upserts to particular (reduced) datoms, i.e., to particular [e a v] triples that caused the constraint violation. Attributing constraint violations to input data is more difficult to the multiple rewriting passes the input undergoes.

TypeDisagreements

A transaction tried to assert a datom or datoms with the wrong value v type(s).

Fields of TypeDisagreements

conflicting_datoms: BTreeMap<(Entid, Entid, TypedValue), ValueType>

The key ([e a v]) has an invalid value v: it is not of the expected value type.

CardinalityConflicts

A transaction tried to assert datoms that don't observe the schema's cardinality constraints.

Fields of CardinalityConflicts

conflicts: Vec<CardinalityConflict>